requirements.txt
notebook>=6.0.3
python-dotenv>=0.12.0
librosa>=0.7.2
boto3>=1.12.28
psycopg2-binary>=2.8.4
matplotlib>=3.2.1
pandas>=1.0.3
numpy>=1.18.2
tqdm>=4.43.0
-e git+https://github.com/detly/gammatone.git#egg=gammatone